Add `RichText` component via `mdast`.
This component uses `mdast` to convert a markdown parse tree into react elements. It's not terribly fancy and it comes with drawbacks, but it should be simple and flexible enough to use for our purposes currently. If it becomes too much of a burden we could always write our own parser.