Remove `csswring` due to mimimizing conflicts.
If `csswring` is used and an empty class is used in a `composes` line lots of errors are generated since `csswring` deletes the class before `css-loader` can work its magic. It turns out `css-loader` includes `cssnano` for minimizing which probably works even better, so just use that and remove another dependency.